% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{coalesce}
\alias{coalesce}
\title{Coalesce values}
\usage{
coalesce(a, b)
}
\arguments{
\item{a}{high priority}

\item{b}{low priority}
}
\value{
return values according to priority, and NA where both a and b are missing.
}
\description{
Coalesce values
}

