% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{count.na}
\alias{count.na}
\title{Count NA}
\usage{
count.na(x)
}
\arguments{
\item{x}{value}
}
\value{
return values according to priority, and NA where both a and b are missing.
}
\description{
Count NA
}

