% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fill.f}
\alias{fill.f}
\title{Replace conditional on given function}
\usage{
fill.f(x, value, f = is.na)
}
\arguments{
\item{x}{values}

\item{value}{replacement value}
}
\description{
Replace conditional on given function
}

