% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten}
\alias{flatten}
\title{Flatten list and keep names}
\usage{
flatten(x, use.names = TRUE, classes = "ANY")
}
\arguments{
\item{x}{List.}

\item{use.names}{Logical. Should the names of x be kept?}

\item{classes}{Character. Vector of class names, or "ANY" to match any class.}
}
\value{
Flattened named list.
}
\description{
Flatten list and keep names
}
\references{
https://stackoverflow.com/questions/49252400/r-purrr-flatten-list-of-named-lists-to-list-and-keep-names
}
