% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kaggle_build_script}
\alias{kaggle_build_script}
\title{Build bash script to execute kaggle commands}
\usage{
kaggle_build_script(command)
}
\arguments{
\item{command}{String with command to execute}
}
\description{
Build bash script to execute kaggle commands
}
