% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kaggle_command_to_df}
\alias{kaggle_command_to_df}
\title{Redirect output of kaggle command to data.frame}
\usage{
kaggle_command_to_df(command)
}
\arguments{
\item{command}{String with command to execute}
}
\description{
Redirect output of kaggle command to data.frame
}
