% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-data.R
\name{kaggle_competitions_data_list}
\alias{kaggle_competitions_data_list}
\title{CompetitionsDataListFiles}
\usage{
kaggle_competitions_data_list(id)
}
\arguments{
\item{id}{string, Competition name. Required: TRUE.}
}
\description{
List competition data files
}
