% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competitions.R
\name{kaggle_competitions_download_files}
\alias{kaggle_competitions_download_files}
\title{Download Competition Files}
\usage{
kaggle_competitions_download_files(
  competition,
  file_name = NULL,
  path = NULL,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{competition}{Competition URL suffix (use `kaggle_competitions_list()` to show options)}

\item{file_name}{File name, all files downloaded if not provided}

\item{path}{Folder where file(s) will be downloaded, defaults to current working directory}

\item{force}{Skip check whether local version of file is up to date, force file download}

\item{quiet}{Suppress printing information about the upload/download progress}
}
\value{
`NULL`
}
\description{
Note: you will need to accept competition rules at https://www.kaggle.com/c/<competition-name>/rules.
}
\examples{
kaggle_competitions_download_files(competition = "favorita-grocery-sales-forecasting")
kaggle_competitions_download_files(competition = "favorita-grocery-sales-forecasting",
                                   file_name = "test.csv.7z")
}
