% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competitions.R
\name{kaggle_competitions_files}
\alias{kaggle_competitions_files}
\title{List Competition Files}
\usage{
kaggle_competitions_files(competition, quiet = FALSE)
}
\arguments{
\item{competition}{Competition URL suffix (use `kaggle_competitions_list()` to show options)}

\item{quiet}{Suppress printing information about the upload/download progress}
}
\description{
Note: you will need to accept competition rules at https://www.kaggle.com/c/<competition-name>/rules.
}
\examples{
kaggle_competitions_files(competition = "favorita-grocery-sales-forecasting")


}
