% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competitions.R
\name{kaggle_competitions_leaderboard}
\alias{kaggle_competitions_leaderboard}
\title{Get Competition Leaderboard}
\usage{
kaggle_competitions_leaderboard(
  competition,
  show = TRUE,
  download = FALSE,
  path = NULL,
  quiet = FALSE
)
}
\arguments{
\item{competition}{Competition URL suffix (use `kaggle_competitions_list()` to show options)}

\item{show}{Show the top of the leaderboard}

\item{download}{Download entire leaderboard}

\item{path}{Folder where file(s) will be downloaded, defaults to current working directory}

\item{quiet}{Suppress printing information about the upload/download progress}
}
\description{
Get Competition Leaderboard
}
\examples{
kaggle_competitions_leaderboard(competition = "favorita-grocery-sales-forecasting")
kaggle_competitions_leaderboard(competition = "favorita-grocery-sales-forecasting",
                                download = TRUE)
kaggle_competitions_leaderboard(competition = "favorita-grocery-sales-forecasting",
                                download = TRUE, path = "./path")
}
