% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-leaderboard.R
\name{kaggle_competitions_leaderboard_view}
\alias{kaggle_competitions_leaderboard_view}
\title{CompetitionViewLeaderboard}
\usage{
kaggle_competitions_leaderboard_view(id)
}
\arguments{
\item{id}{string, Competition name. Required: TRUE.}
}
\description{
VIew competition leaderboard
}
