% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions.R
\name{kaggle_competitions_list}
\alias{kaggle_competitions_list}
\title{CompetitionsList}
\usage{
kaggle_competitions_list(page = 1, search = "")
}
\arguments{
\item{page}{integer, Page number. Defaults to 1. Required: FALSE.}

\item{search}{string, Search terms. Defaults to . Required: FALSE.}
}
\description{
List competitions
}
