% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competitions.R
\name{kaggle_competitions_submit}
\alias{kaggle_competitions_submit}
\title{Submit to a Competition}
\usage{
kaggle_competitions_submit(
  file_name,
  message,
  competition = NULL,
  quiet = FALSE
)
}
\arguments{
\item{file_name}{File for upload (full path)}

\item{message}{Message describing this submission}

\item{competition}{Competition URL suffix (use `kaggle_competitions_list()` to show options). If empty, the default competition will be used (use "kaggle config set competition")"}

\item{quiet}{Suppress printing information about the upload/download progress}
}
\description{
Note: you will need to accept competition rules at https://www.kaggle.com/c/<competition-name>/rules.
}
\examples{
\dontrun{
kaggle_competitions_submit(competition = "favorita-grocery-sales-forecasting",
                           file_name = "sample_submission_favorita.csv.7z",
                           message = "My submission message")
}


}
