% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{kaggle_config_set}
\alias{kaggle_config_set}
\title{Set a Configuration Value}
\usage{
kaggle_config_set(name, value)
}
\arguments{
\item{name}{Name of the configuration parameter (one of competition, path, proxy)}

\item{value}{Value of the configuration parameter}
}
\description{
Valid values depending on name
- competition: Competition URL suffix (use `kaggle_competitions_list()` to show options)
- path: Folder where file(s) will be downloaded, defaults to current working directory
- proxy: Proxy for HTTP requests
}
\examples{
kaggle_config_set(name = "competition", value = "titanic")


}
