% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{kaggle_config_unset}
\alias{kaggle_config_unset}
\title{Clear a configuration value}
\usage{
kaggle_config_unset(name)
}
\arguments{
\item{name}{Name of the configuration parameter}
}
\description{
Clear a configuration value
}
\examples{
kaggle_config_unset(name = "competition")


}
