% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{kaggle_datasets_create}
\alias{kaggle_datasets_create}
\title{Create a new dataset}
\usage{
kaggle_datasets_create(
  folder,
  dir_mode = c("skip", "zip", "tar"),
  public = FALSE,
  keep_tabular = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{folder}{Folder for upload, containing data files and a special dataset-metadata.json file (https://github.com/Kaggle/kaggle-api/wiki/Dataset-Metadata).}

\item{dir_mode}{What to do with directories (default is "skip"): "skip" - ignore; "zip" - compressed upload; "tar" - uncompressed upload}

\item{public}{Create publicly (default is private)}

\item{keep_tabular}{Do not convert tabular files to CSV (default is to convert)}

\item{quiet}{Suppress printing information about the upload/download progress}
}
\description{
Create a new dataset
}
\examples{
kaggle_datasets_create(folder = "./path/to/dataset")


}
