% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kaggle_datasets_create_new}
\alias{kaggle_datasets_create_new}
\title{DatasetsCreateNew}
\usage{
kaggle_datasets_create_new(datasetNewRequest)
}
\arguments{
\item{datasetNewRequest}{Information for creating a new dataset. Required: TRUE.}
}
\description{
Create a new dataset
}
