% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kaggle_datasets_create_version}
\alias{kaggle_datasets_create_version}
\title{DatasetsCreateVersion}
\usage{
kaggle_datasets_create_version(ownerSlug, datasetSlug, datasetNewVersionRequest)
}
\arguments{
\item{ownerSlug}{string, Dataset owner. Required: TRUE.}

\item{datasetSlug}{string, Dataset name. Required: TRUE.}

\item{datasetNewVersionRequest}{Information for creating a new dataset version.
Required: TRUE.}
}
\description{
Create a new dataset version
}
