% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{kaggle_datasets_download}
\alias{kaggle_datasets_download}
\title{Download dataset files}
\usage{
kaggle_datasets_download(
  dataset,
  file_name = NULL,
  path = NULL,
  unzip = FALSE,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{dataset}{Dataset URL suffix in format <owner>/<dataset-name> (use "kaggle_datasets_list()" to show options)}

\item{file_name}{File name, all files downloaded if not provided (use kaggle_datasets_files(dataset = "<owner>/<dataset-name>") to show options)}

\item{path}{Folder where file(s) will be downloaded, defaults to current working directory}

\item{unzip}{Unzip the downloaded file. Will delete the zip file when completed.}

\item{force}{Skip check whether local version of file is up to date, force file download}

\item{quiet}{Suppress printing information about the upload/download progress}
}
\description{
Please note that BigQuery datasets cannot be downloaded.
}
\examples{
kaggle_datasets_download("zillow/zecon")
kaggle_datasets_download("zillow/zecon", file_name = "State_time_series.csv")


}
