% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{kaggle_datasets_files}
\alias{kaggle_datasets_files}
\title{List Files for a Dataset}
\usage{
kaggle_datasets_files(dataset)
}
\arguments{
\item{dataset}{Dataset URL suffix in format <owner>/<dataset-name> (use "kaggle_datasets_list()" to show options)}
}
\description{
List Files for a Dataset
}
\examples{
kaggle_datasets_files("zillow/zecon")
}
