% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{kaggle_datasets_init}
\alias{kaggle_datasets_init}
\title{Initialize Metadata file for Dataset Creation}
\usage{
kaggle_datasets_init(folder)
}
\arguments{
\item{folder}{Folder for upload, containing data files and a special dataset-metadata.json file (https://github.com/Kaggle/kaggle-api/wiki/Dataset-Metadata).}
}
\description{
Initialize Metadata file for Dataset Creation
}
\examples{
kaggle_datasets_init(folder = "./path/to/dataset")


}
