% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kaggle_datasets_list}
\alias{kaggle_datasets_list}
\title{DatasetsList}
\usage{
kaggle_datasets_list(page = 1, search = "", ownerSlug = NULL,
  datasetSlug = NULL)
}
\arguments{
\item{page}{integer, Page number. Defaults to 1. Retrieve datasets via
page, search, or (ownerSlug and datasetSlug)}

\item{search}{string, Search terms. Defaults to . Retrieve datasets via
page, search, or (ownerSlug and datasetSlug)}

\item{ownerSlug}{string, Dataset owner. Retrieve datasets via
page, search, or (ownerSlug and datasetSlug). If ownerSlug and datasetSlug
are not NULL, this function looks for the matching dataset.}

\item{datasetSlug}{string, Dataset name. Retrieve datasets via
page, search, or (ownerSlug and datasetSlug). If ownerSlug and datasetSlug
are not NULL, this function looks for the matching dataset.}
}
\description{
List datasets
}
