% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{kaggle_datasets_metadata}
\alias{kaggle_datasets_metadata}
\title{Download Metadata for an Existing Dataset}
\usage{
kaggle_datasets_metadata(dataset, path = NULL)
}
\arguments{
\item{dataset}{Dataset URL suffix in format <owner>/<dataset-name> (use `kaggle_datasets_list()`to show options)}

\item{path}{Location to download dataset metadata to. Defaults to current working directory}
}
\description{
Download Metadata for an Existing Dataset
}
\examples{
kaggle_datasets_metadata(dataset = "zillow/zecon", path = "./path/to/download")
}
