% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{kaggle_datasets_status}
\alias{kaggle_datasets_status}
\title{Get Dataset Creation Status}
\usage{
kaggle_datasets_status(dataset)
}
\arguments{
\item{dataset}{dataset Dataset URL suffix in format <owner>/<dataset-name> (use `kaggle_datasets_list()`to show options)}
}
\description{
Get Dataset Creation Status
}
\examples{
kaggle_datasets_status("zillow/zecon")
}
