% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{kaggle_datasets_version}
\alias{kaggle_datasets_version}
\title{Creata a New Dataset Version}
\usage{
kaggle_datasets_version(
  folder,
  version_notes,
  keep_tabular = FALSE,
  dir_mode = c("skip", "zip", "tar"),
  delete_old_versions = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{folder}{Folder for upload, containing data files and a special dataset-metadata.json file (https://github.com/Kaggle/kaggle-api/wiki/Dataset-Metadata).}

\item{version_notes}{Message describing the new version}

\item{keep_tabular}{Do not convert tabular files to CSV (default is to convert)}

\item{dir_mode}{What to do with directories: "skip" - ignore; "zip" - compressed upload; "tar" - uncompressed upload}

\item{delete_old_versions}{Delete old versions of this dataset}

\item{quiet}{Suppress printing information about the upload/download progress}
}
\description{
Creata a New Dataset Version
}
\examples{
kaggle_datasets_version(folder = "./path/to/dataset", version_notes = "Updated data")
}
