% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kaggle_datasets_view}
\alias{kaggle_datasets_view}
\title{DatasetsView}
\usage{
kaggle_datasets_view(ownerSlug, datasetSlug)
}
\arguments{
\item{ownerSlug}{string, Dataset owner. Required: TRUE.}

\item{datasetSlug}{string, Dataset name. Required: TRUE.}
}
\description{
Show details about a dataset
}
