% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{kaggle_kernels_init}
\alias{kaggle_kernels_init}
\title{Initialize Metadata File for a Kernel}
\usage{
kaggle_kernels_init(folder)
}
\arguments{
\item{folder}{Folder for upload, containing data files and a special kernel-metadata.json file (https://github.com/Kaggle/kaggle-api/wiki/Kernel-Metadata).}
}
\description{
Initialize Metadata File for a Kernel
}
\examples{
kaggle_kernels_init(folder = "./path/to/kernel")
}
