% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{kaggle_kernels_list}
\alias{kaggle_kernels_list}
\title{List Kernels}
\usage{
kaggle_kernels_list(
  mine = FALSE,
  size = 20,
  search = NULL,
  parent = NULL,
  competition = NULL,
  dataset = NULL,
  user = NULL,
  language = c("all", "python", "r", "sqlite", "julia"),
  kernel_type = c("all", "script", "notebook"),
  output_type = c("all", "visualizations", "data"),
  sort_by = c("hotness", "commentCount", "dateCreated", "dateRun", "relevance",
    "scoreAscending", "scoreDescending", "viewCount", "voteCount")
)
}
\arguments{
\item{mine}{Display only my items}

\item{size}{Number of kernels to list. Default size is 20, max is 100}

\item{search}{Term(s) to search for}

\item{parent}{Find children of the specified parent kernel}

\item{competition}{Find kernels for a given competition}

\item{dataset}{Find kernels for a given dataset}

\item{user}{Find kernels created by a given user}

\item{language}{Specify the language the kernel is written in. Default is 'all'. Valid options are 'all', 'python', 'r', 'sqlite', and 'julia'}

\item{kernel_type}{Specify the type of kernel. Default is 'all'. Valid options are 'all', 'script', and 'notebook'}

\item{output_type}{Search for specific kernel output types. Default is 'all'. Valid options are 'all', 'visualizations', and 'data'}

\item{sort_by}{Sort list results. Default is 'hotness'.  Valid options are 'hotness', 'commentCount', 'dateCreated', 'dateRun', 'relevance', 'scoreAscending', 'scoreDescending', 'viewCount', and 'voteCount'. 'relevance' is only applicable if a search term is specified.}
}
\description{
List Kernels
}
\examples{
kaggle_kernels_list(search = "titanic")
kaggle_kernels_list(language = "python")


}
