% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{kaggle_kernels_output}
\alias{kaggle_kernels_output}
\title{Retrieve a Kernel's output}
\usage{
kaggle_kernels_output(kernel, path = NULL, force = FALSE, quiet = FALSE)
}
\arguments{
\item{kernel}{Kernel URL suffix in format <owner>/<kernel-name> (use `kaggle_kernels_list()` to show options)}

\item{path}{Folder where file(s) will be downloaded, defaults to current working directory}

\item{force}{Skip check whether local version of file is up to date, force file download}

\item{quiet}{Suppress printing information about the upload/download progress}
}
\description{
Retrieve a Kernel's output
}
\examples{
\dontrun{
kaggle_kernels_output(kernel = "mrisdal/exploring-survival-on-the-titanic", path = "/path/to/dest")
}


}
