% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{kaggle_kernels_pull}
\alias{kaggle_kernels_pull}
\title{Pull a Kernel}
\usage{
kaggle_kernels_pull(kernel, path = NULL, metadata = FALSE)
}
\arguments{
\item{kernel}{Kernel URL suffix in format <owner>/<kernel-name> (use `kaggle_kernels_list()` to show options)}

\item{path}{Folder where file(s) will be downloaded, defaults to current working directory}

\item{metadata}{Generate metadata when pulling kernel}
}
\description{
Pull a Kernel
}
\examples{
kaggle_kernels_pull(kernel = "rtatman/list-of-5-day-challenges", path = "/path/to/dest")


}
