% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{kaggle_kernels_push}
\alias{kaggle_kernels_push}
\title{Push a Kernel}
\usage{
kaggle_kernels_push(folder)
}
\arguments{
\item{folder}{Folder for upload, containing data files and a special kernel-metadata.json file (https://github.com/Kaggle/kaggle-api/wiki/Kernel-Metadata).}
}
\description{
Push a Kernel
}
\examples{
kaggle_kernels_push(folder = "./path/to/kernel")
}
