% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggler-package.R
\docType{package}
\name{kaggler-package}
\alias{kaggler}
\alias{kaggler-package}
\title{kaggle api client}
\description{
An interface for interacting with Kaggle's API
}
\seealso{
\href{https://www.kaggle.com}{Kaggle.com}
}
\author{
\strong{Maintainer}: Michael Wayne Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
\keyword{internal}
