% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kgl_as_tbl}
\alias{kgl_as_tbl}
\title{Convert kaggle output to tibble}
\usage{
kgl_as_tbl(x)
}
\arguments{
\item{x}{Output from kaggle function}
}
\value{
Print out of summary info and a tibble of the data.
}
\description{
Convert kaggle response objects into informative tibbles
}
