% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{kgl_auth}
\alias{kgl_auth}
\title{Kaggle API authorization}
\usage{
kgl_auth(username = NULL, key = NULL, creds_file = NULL)
}
\arguments{
\item{username}{Name associated with user's account at
\href{https://www.kaggle.com}{Kaggle.com}.}

\item{key}{Key (a string of random letters and numbers) generated by Kaggle. Users can
obtain or lookup their keys by signing into Kaggle and clicking on 'Account" or by
substituting their username at kaggle.com/{username}/account.}

\item{creds_file}{Path to user's \code{kaggle.json} file. See @details for
more information.}
}
\description{
Authorizing access to Kaggle's API
}
\details{
When Kaggle generates a key, it's sent to users as a
\code{kaggle.json} file, which is likely saved in a place like the default
Downloads folder or something. Users can (a) save the \code{kaggle.json}
file to a \code{.kaggle} directory in the user's machine's home folder
(i.e.,  \code{~/.kaggle/kaggle.json}), (b) specify the path to their
\code{kaggle.json} file using this \code{creds_file} argument, or (c)
save their \code{kaggle.json} file in their current working directory. The
third options (options c) is the least desirable as your working directory
will presumably change between sessions where you'd still want to have
authorized access to Kaggle's API. Either of the first two options would be
a good choice.

Kaggle encourages users to save the "kaggle.json" credentials file in their
.kaggle file in their home directory (\code{~/.kaggle/kaggle.json}). If all
arguments are left NULL (their defaults), this function will look to see if
you computer has a \code{~/.kaggle/kaggle.json} file. If it can't find it
there, it'll look in the current working directory. If it finds the file,
it'll use it. Otherwise it'll look for environment variables, which it will
create and save for you manually the first time you enter your username/key
or path to your \code{kaggle.json} file.
}
