% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{kgl_auth_file_setup}
\alias{kgl_auth_file_setup}
\title{Setup API authorization file}
\usage{
kgl_auth_file_setup(path)
}
\arguments{
\item{path}{Character. Path to \code{kaggle.json}}
}
\value{
Nothing.
}
\description{
This function aims to simplify the authorization setup process. Once the API key has been generated and the \code{kaggle.json} file is saved, use the path of the downloaded file to have it correctly stored in the home directory. This function will only need to be ran once per API key that is generated.
}
\examples{
\dontrun{
library(kaggler)

kgl_auth_file_setup("/Users/Kow/Downloads/kaggle.json")
}
}
\concept{Authorization}
