% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-submissions.R
\name{kgl_competition_submit}
\alias{kgl_competition_submit}
\title{Submit data to competition}
\usage{
kgl_competition_submit(id, file, submission_description = "")
}
\arguments{
\item{id}{Character.}

\item{file}{Character. Competition submission file name.}

\item{submission_description}{Character. Description of the submission.}
}
\value{
Nothing.
}
\description{
Submit data to competition
}
\examples{
\dontrun{
kgl_competition_submit(
  id = "titanic",
  file = "gender_submission.csv",
  submission_description = "Trained via rpart"
)
}
}
\seealso{
Other Competitions: 
\code{\link{kgl_competitions_data_download_all}()},
\code{\link{kgl_competitions_data_download}()},
\code{\link{kgl_competitions_data_list}()},
\code{\link{kgl_competitions_leaderboard_download}()},
\code{\link{kgl_competitions_leaderboard_view}()},
\code{\link{kgl_competitions_list_submissions}()},
\code{\link{kgl_competitions_list}()}
}
\concept{Competitions}
