% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-data.R
\name{kgl_competitions_data_download}
\alias{kgl_competitions_data_download}
\title{CompetitionsDataDownloadFile}
\usage{
kgl_competitions_data_download(id, fileName)
}
\arguments{
\item{id}{string, Competition name. Required: TRUE.}

\item{fileName}{string, Competition name. Required: TRUE.}
}
\description{
Download competition data file
}
