% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-leaderboard.R
\name{kgl_competitions_leaderboard_download}
\alias{kgl_competitions_leaderboard_download}
\title{Competition Download Leaderboard}
\usage{
kgl_competitions_leaderboard_download(id, output_dir = ".")
}
\arguments{
\item{id}{Character. Competition name.}

\item{output_dir}{Character. Directory to save the file.}
}
\value{
An invisible \code{\link[httr:GET]{httr::GET()}} object.
}
\description{
Download the entire competition leaderboard into a zip file.
}
\seealso{
Other Competitions: 
\code{\link{kgl_competitions_data_download_all}()},
\code{\link{kgl_competitions_data_download}()},
\code{\link{kgl_competitions_data_list}()},
\code{\link{kgl_competitions_leaderboard_view}()},
\code{\link{kgl_competitions_list}()}
}
\concept{Competitions}
