% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-submissions.R
\name{kgl_competitions_submissions_submit}
\alias{kgl_competitions_submissions_submit}
\title{CompetitionsSubmissionsSubmit}
\usage{
kgl_competitions_submissions_submit(blobFileTokens, submissionDescription, id)
}
\arguments{
\item{blobFileTokens}{string, Token identifying location of uploaded submission
file. Required: TRUE.}

\item{submissionDescription}{string, Description of competition submission.
Required: TRUE.}

\item{id}{string, Competition name. Required: TRUE.}
}
\description{
Submit to competition
}
