% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-submissions.R
\name{kgl_competitions_submissions_upload}
\alias{kgl_competitions_submissions_upload}
\title{CompetitionsSubmissionsUpload}
\usage{
kgl_competitions_submissions_upload(file, guid, contentLength,
  lastModifiedDateUtc)
}
\arguments{
\item{file}{file, Competition submission file. Required: TRUE.}

\item{guid}{string, Location where submission should be uploaded. Required:
TRUE.}

\item{contentLength}{integer, Content length of file in bytes. Required: TRUE.}

\item{lastModifiedDateUtc}{integer, Last modified date of file in milliseconds
since epoch in UTC. Required: TRUE.}
}
\description{
Upload competition submission file
}
