% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-submissions.R
\name{kgl_competitions_submissions_url}
\alias{kgl_competitions_submissions_url}
\title{CompetitionsSubmissionsUrl}
\usage{
kgl_competitions_submissions_url(fileName = NULL, contentLength,
  lastModifiedDateUtc)
}
\arguments{
\item{fileName}{string, Competition submission file name. Required: FALSE.}

\item{contentLength}{integer, Content length of file in bytes. Required: TRUE.}

\item{lastModifiedDateUtc}{integer, Last modified date of file in milliseconds
since epoch in UTC. Required: TRUE.}
}
\description{
Generate competition submission URL
}
