% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kgl_datasets_create_new}
\alias{kgl_datasets_create_new}
\title{DatasetsCreateNew}
\usage{
kgl_datasets_create_new(datasetNewRequest)
}
\arguments{
\item{datasetNewRequest}{Information for creating a new dataset. Required: TRUE.}
}
\description{
Create a new dataset
}
\seealso{
Other Datasets: 
\code{\link{kgl_datasets_create_version}()},
\code{\link{kgl_datasets_download}()},
\code{\link{kgl_datasets_list}()},
\code{\link{kgl_datasets_upload_file}()},
\code{\link{kgl_datasets_view}()}
}
\concept{Datasets}
