% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kgl_datasets_download}
\alias{kgl_datasets_download}
\title{DatasetsDownloadFile}
\usage{
kgl_datasets_download(ownerSlug, datasetSlug, fileName,
  datasetVersionNumber = NULL)
}
\arguments{
\item{ownerSlug}{string, Dataset owner. Required: TRUE.}

\item{datasetSlug}{string, Dataset name. Required: TRUE.}

\item{fileName}{string, File name. Required: TRUE.}

\item{datasetVersionNumber}{string, Dataset version number. Required: FALSE.}
}
\description{
Download dataset file
}
