% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kgl_datasets_list}
\alias{kgl_datasets_list}
\title{DatasetsList}
\usage{
kgl_datasets_list(page = 1, search = NULL, owner_dataset = NULL)
}
\arguments{
\item{page}{Numeric. Page number. Defaults to 1. Retrieve datasets via page, search, or (ownerSlug and datasetSlug)}

\item{search}{Character. Search terms. Defaults to . Retrieve datasets via page, search, or (ownerSlug and datasetSlug)}

\item{owner_dataset}{Character. Alternative to page/search.  The owner and dataset slug as it appears in the URL, i.e., \code{"mathan/fifa-2018-match-statistics"}.}
}
\description{
List datasets
}
\seealso{
Other Datasets: 
\code{\link{kgl_datasets_create_new}()},
\code{\link{kgl_datasets_create_version}()},
\code{\link{kgl_datasets_download}()},
\code{\link{kgl_datasets_upload_file}()},
\code{\link{kgl_datasets_view}()}
}
\concept{Datasets}
