% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgl_flow_meta.R
\name{kgl_flow_competition_info}
\alias{kgl_flow_competition_info}
\title{Show competition info}
\usage{
kgl_flow_competition_info()
}
\value{
Compeition info.
}
\description{
This function will show some competition information for the current project.
}
\examples{
\dontrun{
kgl_flow("titanic")

kgl_flow_competition_info()
}
}
\seealso{
Other Kaggle Flow: 
\code{\link{kgl_flow_leaderboard}()},
\code{\link{kgl_flow_load}()},
\code{\link{kgl_flow_meta}()},
\code{\link{kgl_flow}()}
}
\concept{Kaggle Flow}
