% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgl_flow_leaderboard.R
\name{kgl_flow_leaderboard}
\alias{kgl_flow_leaderboard}
\title{Download the Competition Leaderboard}
\usage{
kgl_flow_leaderboard()
}
\value{
A data frame with one row per user who has submitted
results to the competition.
}
\description{
Download the entire leaderboard for the competition specified
in the metadata.
}
\examples{
\dontrun{
library(kaggler)

kgl_flow("titanic")

kgl_flow_leaderboard()
}
}
\seealso{
Other Kaggle Flows: 
\code{\link{kgl_flow_competition_info}()},
\code{\link{kgl_flow_load}()},
\code{\link{kgl_flow_meta}()},
\code{\link{kgl_flow}()}
}
\concept{Kaggle Flows}
