% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgl_flow_meta.R
\name{kgl_flow_meta}
\alias{kgl_flow_meta}
\title{Read the project's metadata}
\usage{
kgl_flow_meta()
}
\value{
A data frame with one row per competition dataset.
}
\description{
Read the metadata of the competitions datasets.
}
\examples{
\dontrun{
library(kaggler)

kgl_flow("titanic")

kgl_flow_meta()
}
}
\seealso{
Other Kaggle Flows: 
\code{\link{kgl_flow_load}()},
\code{\link{kgl_flow}()}
}
\concept{Kaggle Flows}
