% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{rmsle}
\alias{rmsle}
\title{Root Mean Squared Log Error}
\usage{
rmsle(y.hat, y)
}
\arguments{
\item{y.hat}{Fitted values}

\item{y}{Realized value}
}
\description{
Root Mean Squared Log Error
}

