% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_chains.R
\name{steady_state}
\alias{steady_state}
\title{\code{steady_state} Steady State Convergence}
\usage{
steady_state(P)
}
\arguments{
\item{P}{A transition probability matrix}
}
\description{
\code{steady_state} Steady State Convergence
}

