% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competitions.R
\name{kaggle_competitions_list}
\alias{kaggle_competitions_list}
\title{List Competitions}
\usage{
kaggle_competitions_list(
  group = c("general", "entered", "inClass"),
  category = c("all", "featured", "research", "recruitment", "gettingStarted",
    "masters", "playground"),
  sort_by = c("latestDeadline", "grouped", "prize", "earliestDeadline",
    "numberOfTeams", "recentlyCreated"),
  search = NULL
)
}
\arguments{
\item{group}{Search for competitions in a specific group. Default is 'general'. Valid options are 'general', 'entered', and 'inClass'}

\item{category}{Search for competitions of a specific category. Default is 'all'. Valid options are 'all', 'featured', 'research', 'recruitment', 'gettingStarted', 'masters', and 'playground'}

\item{sort_by}{Sort list results. Default is 'latestDeadline'. Valid options are 'grouped', 'prize', 'earliestDeadline', 'latestDeadline', 'numberOfTeams', and 'recentlyCreated'}

\item{search}{Term(s) to search for}
}
\value{
`data.frame`
}
\description{
List Competitions
}
\examples{
kaggle_competitions_list(category = "gettingStarted")
kaggle_competitions_list(search = "health")


}
