% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{kaggle_datasets_list}
\alias{kaggle_datasets_list}
\title{List Datasets}
\usage{
kaggle_datasets_list(
  sort_by = c("hottest", "votes", "updated", "active"),
  file_type = c("all", "csv", "sqlite", "json", "bigQuery"),
  license_name = c("all", "cc", "gpl", "odb", "other"),
  max_size = NULL,
  min_size = NULL,
  tag_ids = NULL,
  search = NULL,
  mine = NULL,
  user = NULL
)
}
\arguments{
\item{sort_by}{Sort list results. Default is 'hottest'. Valid options are 'hottest', 'votes', 'updated', and 'active'}

\item{file_type}{Search for datasets with a specific file type. Default is 'all'. Valid options are 'all', 'csv', 'sqlite', 'json', and 'bigQuery'. Please note that bigQuery datasets cannot be downloaded}

\item{license_name}{Search for datasets with a specific license. Default is 'all'. Valid options are 'all', 'cc', 'gpl', 'odb', and 'other'}

\item{max_size}{Specify the maximum size of the dataset to return (bytes)}

\item{min_size}{Specify the minimum size of the dataset to return (bytes)}

\item{tag_ids}{Search for datasets that have specific tags. Tag list should be comma separated}

\item{search}{Term(s) to search for}

\item{mine}{Display only my items}

\item{user}{Find public datasets owned by a specific user or organization}
}
\description{
List Datasets
}
\examples{
kaggle_datasets_list(search = "demographies")
kaggle_datasets_list(sort_by = "votes")
}
