% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-data.R
\name{kgl_competitions_data_download}
\alias{kgl_competitions_data_download}
\title{CompetitionsDataDownloadFile}
\usage{
kgl_competitions_data_download(id, file_name, output_dir = ".")
}
\arguments{
\item{id}{Character. Competition name.}

\item{file_name}{Character. Competition name.}

\item{output_dir}{Character. Directory to save the file.}
}
\value{
An invisible \code{\link[httr:GET]{httr::GET()}} object.
}
\description{
Download competition data file
}
\seealso{
Other Competitions: 
\code{\link{kgl_competitions_data_download_all}()},
\code{\link{kgl_competitions_data_list}()},
\code{\link{kgl_competitions_leaderboard_download}()},
\code{\link{kgl_competitions_leaderboard_view}()},
\code{\link{kgl_competitions_list}()}
}
\concept{Competitions}
