% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-data.R
\name{kgl_competitions_data_list}
\alias{kgl_competitions_data_list}
\title{Competitions Data List Files}
\usage{
kgl_competitions_data_list(id, clean_response = TRUE)
}
\arguments{
\item{id}{Character. Competition name.}

\item{clean_response}{Logical. Clean the response from the Kaggle API. If \code{FALSE}, this will return the object from the \code{\link[httr:GET]{httr::GET()}} call.}
}
\value{
Based on \code{clean_response}: A tibble containing information on the given \code{id} or a \code{\link[httr:GET]{httr::GET()}} object.
}
\description{
List competition data files
}
\seealso{
Other Competitions: 
\code{\link{kgl_competitions_data_download_all}()},
\code{\link{kgl_competitions_data_download}()},
\code{\link{kgl_competitions_leaderboard_download}()},
\code{\link{kgl_competitions_leaderboard_view}()},
\code{\link{kgl_competitions_list}()}
}
\concept{Competitions}
