% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-leaderboard.R
\name{kgl_competitions_leaderboard_download}
\alias{kgl_competitions_leaderboard_download}
\title{Competition Download Leaderboard}
\usage{
kgl_competitions_leaderboard_download(
  id,
  output_dir = ".",
  clean_response = TRUE
)
}
\arguments{
\item{id}{Character. Competition name.}

\item{output_dir}{Character. Directory to save the file.}

\item{clean_response}{Logical. Download the data within the response from the Kaggle API. If \code{FALSE}, this will return the object from the \code{\link[httr2:req_perform]{httr2::req_perform()}} call.}
}
\value{
An invisible \code{\link[httr:GET]{httr::GET()}} object.
}
\description{
Download the entire competition leaderboard into a zip file.
}
\seealso{
Other Competitions: 
\code{\link{kgl_competition_submit}()},
\code{\link{kgl_competitions_data_download_all}()},
\code{\link{kgl_competitions_data_download}()},
\code{\link{kgl_competitions_data_list}()},
\code{\link{kgl_competitions_leaderboard_view}()},
\code{\link{kgl_competitions_list_submissions}()},
\code{\link{kgl_competitions_list}()}
}
\concept{Competitions}
