% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-leaderboard.R
\name{kgl_competitions_leaderboard_download}
\alias{kgl_competitions_leaderboard_download}
\title{CompetitionDownloadLeaderboard}
\usage{
kgl_competitions_leaderboard_download(id)
}
\arguments{
\item{id}{string, Competition name. Required: TRUE.}
}
\description{
Download competition leaderboard
}
