% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-leaderboard.R
\name{kgl_competitions_leaderboard_view}
\alias{kgl_competitions_leaderboard_view}
\title{Competition View Leaderboard}
\usage{
kgl_competitions_leaderboard_view(id)
}
\arguments{
\item{id}{Character. Competition name. Required: TRUE.}
}
\description{
VIew competition leaderboard
}
