% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions-leaderboard.R
\name{kgl_competitions_leaderboard_view}
\alias{kgl_competitions_leaderboard_view}
\title{Competition view leaderboard}
\usage{
kgl_competitions_leaderboard_view(id, clean_response = TRUE)
}
\arguments{
\item{id}{Character. Competition name.}

\item{clean_response}{Logical. Clean the response from the Kaggle API. If \code{FALSE}, this will return the object from the \code{\link[httr2:req_perform]{httr2::req_perform()}} call.}
}
\value{
Based on \code{clean_response}: A tibble containing information on the given \code{id} or a \code{\link[httr:GET]{httr::GET()}} object.
}
\description{
View the top 50 positions in a competition's leaderboard.
}
\seealso{
Other Competitions: 
\code{\link{kgl_competitions_data_download_all}()},
\code{\link{kgl_competitions_data_download}()},
\code{\link{kgl_competitions_data_list}()},
\code{\link{kgl_competitions_leaderboard_download}()},
\code{\link{kgl_competitions_list}()}
}
\concept{Competitions}
