% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions.R
\name{kgl_competitions_list}
\alias{kgl_competitions_list}
\title{Competitions List}
\usage{
kgl_competitions_list(page = 1, search = NULL)
}
\arguments{
\item{page}{integer, Page number. Defaults to 1. Required: FALSE.}

\item{search}{string, Search terms. Required: FALSE.}
}
\description{
List competitions
}
