% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-submissions.R
\name{kgl_competitions_submissions_submit}
\alias{kgl_competitions_submissions_submit}
\title{CompetitionsSubmissionsSubmit}
\usage{
kgl_competitions_submissions_submit(
  blob_file_tokens,
  submission_description,
  id
)
}
\arguments{
\item{blob_file_tokens}{string, Token identifying location of uploaded submission
file. Required: TRUE.}

\item{submission_description}{string, Description of competition submission.
Required: TRUE.}

\item{id}{string, Competition name. Required: TRUE.}
}
\description{
Submit to competition
}
