% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-submissions.R
\name{kgl_competitions_submissions_submit}
\alias{kgl_competitions_submissions_submit}
\title{Competition Submission Submit}
\usage{
kgl_competitions_submissions_submit(
  blob_file_tokens,
  submission_description,
  id
)
}
\arguments{
\item{blob_file_tokens}{Character. Token identifying location of uploaded submission file. Required: TRUE.}

\item{submission_description}{Character. Description of competition submission. Required: TRUE.}

\item{id}{Character. Competition name. Required: TRUE.}
}
\description{
Submit to competition
}
