% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-submissions.R
\name{kgl_competitions_submissions_upload}
\alias{kgl_competitions_submissions_upload}
\title{Competition Submission Upload}
\usage{
kgl_competitions_submissions_upload(file, guid)
}
\arguments{
\item{file}{file, Competition submission file. Required: TRUE.}

\item{guid}{Character. Location where submission should be uploaded. Required: TRUE.}
}
\description{
Upload competition submission file
}
