% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-submissions.R
\name{kgl_competitions_submissions_url}
\alias{kgl_competitions_submissions_url}
\title{CompetitionsSubmissionsUrl}
\usage{
kgl_competitions_submissions_url(file_name = NULL)
}
\arguments{
\item{file_name}{string, Competition submission file name. Required: FALSE.
since epoch in UTC. Required: TRUE.}
}
\description{
Generate competition submission URL
}
