% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kgl_datasets_create_version}
\alias{kgl_datasets_create_version}
\title{DatasetsCreateVersion}
\usage{
kgl_datasets_create_version(owner_dataset, datasetNewVersionRequest)
}
\arguments{
\item{owner_dataset}{The owner and data set slug as it appears in the URL,
i.e., \code{"mathan/fifa-2018-match-statistics"}.}

\item{datasetNewVersionRequest}{Information for creating a new dataset version.
Required: TRUE.}
}
\description{
Create a new dataset version
}
\seealso{
Other Datasets: 
\code{\link{kgl_datasets_create_new}()},
\code{\link{kgl_datasets_download}()},
\code{\link{kgl_datasets_list}()},
\code{\link{kgl_datasets_upload_file}()},
\code{\link{kgl_datasets_view}()}
}
\concept{Datasets}
