% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kgl_datasets_download}
\alias{kgl_datasets_download}
\title{DatasetsDownloadFile}
\usage{
kgl_datasets_download(owner_dataset, fileName, datasetVersionNumber = NULL)
}
\arguments{
\item{owner_dataset}{The owner and data set slug as it appears in the URL,
i.e., \code{"mathan/fifa-2018-match-statistics"}.}

\item{fileName}{string, File name. Required: TRUE.}

\item{datasetVersionNumber}{string, Dataset version number. Required: FALSE.}
}
\description{
Download dataset file
}
