% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kgl_datasets_list}
\alias{kgl_datasets_list}
\title{DatasetsList}
\usage{
kgl_datasets_list(page = 1, search = "", owner_dataset = NULL)
}
\arguments{
\item{page}{integer, Page number. Defaults to 1. Retrieve datasets via
page, search, or (ownerSlug and datasetSlug)}

\item{search}{string, Search terms. Defaults to . Retrieve datasets via
page, search, or (ownerSlug and datasetSlug)}

\item{owner_dataset}{Alternative to page/search.  The owner and dataset
slug as it appears in the URL, i.e.,
\code{"mathan/fifa-2018-match-statistics"}.}
}
\description{
List datasets
}
