% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{kgl_datasets_upload_file}
\alias{kgl_datasets_upload_file}
\title{DatasetsUploadFile}
\usage{
kgl_datasets_upload_file(fileName, contentLength, lastModifiedDateUtc)
}
\arguments{
\item{fileName}{string, Dataset file name. Required: TRUE.}

\item{contentLength}{integer, Content length of file in bytes. Required: TRUE.}

\item{lastModifiedDateUtc}{integer, Last modified date of file in milliseconds
since epoch in UTC. Required: TRUE.}
}
\description{
Get URL and token to start uploading a data file
}
