% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgl_flow_load.R
\name{kgl_flow_load}
\alias{kgl_flow_load}
\title{Loan the Kaggle data}
\usage{
kgl_flow_load(envir = parent.frame())
}
\arguments{
\item{envir}{Environment to put the loaded kaggle data.}
}
\value{
Nothing.
}
\description{
If all of the files in \verb{_kaggle_data} are in \strong{csv} format (excluding the meta directory) then this function will load all of the kaggle data into the current environment (or the environment of your choosing) via \code{readr::read_csv()}.
}
\examples{
\dontrun{
library(kaggler)

kgl_flow("titanic")

kgl_flow_load()
}
}
\seealso{
Other Kaggle Flows: 
\code{\link{kgl_flow_meta}()},
\code{\link{kgl_flow}()}
}
\concept{Kaggle Flows}
